treewide: allow dependency selection in a cover letter
A lot of series depend on other bits that are in motion, so provide a
way for such dependencies to be specificity in a cover letter.
The expected format is a line containing "depends on:" followed by lore
links to the dependencies. Eg:
depends on:

The regex "(?:based.on|depends.on|depend(?:a|e)ncies):\n(?:\S+@\S+\n)*" is
used to match the line and subsequent links.*
Any such matching is best-effort only, and really patchsets should be
based on the fixes, master or for-next branches of the riscv tree.

*support for "based-on" also includes the "patchew" git trailer type.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>