rv: make the alphanumeric select test report what it complained about
Qinglin pointed out that the output from this test really isn't helpful
at all as it doesn't say *what* it dislikes.
This happens because the diff goes to stdout, but nipa only collects
stderr. Redirect the output there instead.

That said, the output isn't very useful.
Apply some formatting to the output so that it suggests the fixup that
it wants to see.

Reported-by: Qinglin Pan <panqinglin2020@iscas.ac.cn>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>