rv: selects: fix error caused by correctly added new content
The script sees any increase as a bad thing, but adding something in the
correct order will trigger that too. I think this used to work
correctly, but is now broken after my last refactor.
Now, actually check the output for > & <, both of which are present in
a side-by-side diff if there is a suggested change.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>