rv: (try to) fix log output for failed builds
When the initial build fails, we go and break out of the test
immediately, and in some cases print the error.
The issue though is that in the cases that we do actually try to print
the error, we use an empty file as $tmpfile_n is not populated until
later.
Add either /a/ file or a /new/ file, depending on the test, that will
hold the output from the initial build & print that out in the basic
failure case. rv32_defconfig builds already have a file collecting this
information, so print that rather than adding something new.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>